move around perk tier code
