+++ src/lib/constants.ts
@@ -464,8 +464,6 @@ export enum PeakTier {
 	Low = 'low'
 }
 
-export const perkTierCache = new Map<string, number>();
-
 export const minionActivityCache: Map<string, ActivityTaskData> = new Map();
 
 export const ParsedCustomEmojiWithGroups = /(?<animated>a?):(?<name>[^:]+):(?<id>\d{17,20})/;
