+++ src/lib/patreon.ts
@@ -6,7 +6,8 @@ import { cacheBadges } from './badges';
 import { BadgesEnum, BitField, Channel, globalConfig, PatronTierID, PerkTier } from './constants';
 import { fetchSponsors, getUserIdFromGithubID } from './http/util';
 import backgroundImages from './minions/data/bankBackgrounds';
-import { getUsersPerkTier, mahojiUserSettingsUpdate } from './MUser';
+import { mahojiUserSettingsUpdate } from './MUser';
+import { getUsersPerkTier } from './perkTiers';
 import { roboChimpUserFetch } from './roboChimp';
 import { prisma } from './settings/prisma';
 import { Patron } from './types';
