+++ src/mahoji/lib/abstracted_commands/alchCommand.ts
@@ -4,7 +4,7 @@ import { Bank } from 'oldschooljs';
 import { SkillsEnum } from 'oldschooljs/dist/constants';
 import { Item } from 'oldschooljs/dist/meta/types';
 
-import { AlchingActivityTaskOptions } from '../../../lib/types/minions';
+import type { AlchingActivityTaskOptions } from '../../../lib/types/minions';
 import { formatDuration, toKMB } from '../../../lib/util';
 import addSubTaskToActivityTask from '../../../lib/util/addSubTaskToActivityTask';
 import { calcMaxTripLength } from '../../../lib/util/calcMaxTripLength';
