+++ tests/unit/getUsersPerkTier.test.ts
@@ -2,7 +2,8 @@ import { Time } from 'e';
 import { describe, expect, test } from 'vitest';
 
 import { BitField, PerkTier } from '../../src/lib/constants';
-import { getUsersPerkTier, MUserClass } from '../../src/lib/MUser';
+import { MUserClass } from '../../src/lib/MUser';
+import { getUsersPerkTier } from '../../src/lib/perkTiers';
 import { mockMUser } from './utils';
 
 describe('getUsersPerkTier', () => {
